improve ACL engine

+++ cron/centAcl.php
@@ -49,6 +49,8 @@ include_once _CENTREON_PATH_ . "/www/class/centreonContactgroup.class.php";
 
 $centreonDbName = $conf_centreon['db'];
 
+$partition = 0;
+
 function programExit($msg)
 {
     echo "[" . date("Y-m-d H:i:s") . "] " . $msg . "\n";
@@ -76,17 +78,6 @@ try {
     
     $metaObj = new CentreonMeta($pearDB);
     $cgObj = new CentreonContactgroup($pearDB);
-    
-    /*
-     * Detect Which DB layer is used
-     */
-    $DBRESULT = $pearDB->query("SELECT * FROM options WHERE `key` LIKE 'broker'");
-    if (PEAR::isError($DBRESULT)) {
-        print "Cannot Get Monitoring Engine";
-        exit(1);
-    }
-    $row = $DBRESULT->fetchRow();
-    $dbLayer = $row["value"];
 
     /*
      * Lock in MySQL
@@ -154,10 +145,15 @@ try {
     /** **********************************************
      * Remove data from old groups (deleted groups)
      */
-    $aclGroupToDelete = "SELECT DISTINCT acl_group_id FROM $centreonDbName.acl_groups WHERE acl_group_activate = '1'";
-    $aclGroupToDelete2 = "SELECT DISTINCT acl_group_id FROM $centreonDbName.acl_res_group_relations";
-    $pearDBO->query("DELETE FROM centreon_acl WHERE group_id NOT IN ($aclGroupToDelete)");
-    $pearDBO->query("DELETE FROM centreon_acl WHERE group_id NOT IN ($aclGroupToDelete2)");
+    if ($partition == 0) {
+        $aclGroupToDelete = "SELECT DISTINCT acl_group_id FROM $centreonDbName.acl_groups WHERE acl_group_activate = '1'";
+        $aclGroupToDelete2 = "SELECT DISTINCT acl_group_id FROM $centreonDbName.acl_res_group_relations";
+        $pearDBO->query("DELETE FROM centreon_acl WHERE group_id NOT IN ($aclGroupToDelete)");
+        $pearDBO->query("DELETE FROM centreon_acl WHERE group_id NOT IN ($aclGroupToDelete2)");
+    } else {
+        // TODO: Develop Partition management.
+        
+    }
 
     /** ***********************************************
      * Check if some ACL have global options for
@@ -223,7 +219,6 @@ try {
      *  if no : go away.
      *  if yes : let's go to build cache and update database
      */
-
     $tabGroups = array();
     $groupStr = "";
     $query = "SELECT DISTINCT acl_groups.acl_group_id, acl_resources.acl_res_id " .
@@ -263,11 +258,9 @@ try {
         $res->free();
         
         $hostCache = array();
-        $hostNameCache = array();
         $DBRESULT = $pearDB->query("SELECT host_id, host_name FROM host WHERE host_register = '1'");
         while ($h = $DBRESULT->fetchRow()) {
             $hostCache[$h["host_id"]] = $h["host_name"];
-            $hostNameCache[$h["host_name"]] = $h["host_id"];
         }
         $DBRESULT->free();
         unset($h);
@@ -290,12 +283,12 @@ try {
          * Get all included Hosts
          */
         $hostIncCache = array();
-        $DBRESULT = $pearDB->query("SELECT host_id, host_name, acl_res_id FROM `host`, `acl_resources_host_relations` WHERE acl_resources_host_relations.host_host_id = host.host_id AND host.host_register = '1'");
+        $DBRESULT = $pearDB->query("SELECT host_id, acl_res_id FROM `host`, `acl_resources_host_relations` WHERE acl_resources_host_relations.host_host_id = host.host_id AND host.host_register = '1'");
         while ($h = $DBRESULT->fetchRow()) {
             if (!isset($hostIncCache[$h["acl_res_id"]])) {
                 $hostIncCache[$h["acl_res_id"]] = array();
             }
-            $hostIncCache[$h["acl_res_id"]][$h["host_id"]] = $h["host_name"];
+            $hostIncCache[$h["acl_res_id"]][$h["host_id"]] = 1;
         }
         $DBRESULT->free();
         
@@ -303,12 +296,12 @@ try {
          * Get all excluded Hosts
          */
         $hostExclCache = array();
-        $DBRESULT = $pearDB->query("SELECT host_id, host_name, acl_res_id FROM `host`, `acl_resources_hostex_relations` WHERE acl_resources_hostex_relations.host_host_id = host.host_id AND host.host_register = '1'");
+        $DBRESULT = $pearDB->query("SELECT host_id, acl_res_id FROM `host`, `acl_resources_hostex_relations` WHERE acl_resources_hostex_relations.host_host_id = host.host_id AND host.host_register = '1'");
         while ($h = $DBRESULT->fetchRow()) {
             if (!isset($hostExclCache[$h["acl_res_id"]])) {
                 $hostExclCache[$h["acl_res_id"]] = array();
             }
-            $hostExclCache[$h["acl_res_id"]][$h["host_id"]] = $h["host_name"];
+            $hostExclCache[$h["acl_res_id"]][$h["host_id"]] = 1;
         }
         $DBRESULT->free();
         
@@ -429,11 +422,13 @@ try {
             $hgResCache[$row['acl_res_id']][] = $row['hg_id'];
         }
         
-        
+        /* 
+         * Prepare Statment */
+        $injectHandler = $pearDBO->prepare("INSERT INTO centreon_acl (host_id, service_id, group_id) VALUES (?, ?, ?)");
+
         /** ***********************************************
          * Begin to build ACL
          */
-        $strBegin = "INSERT INTO `centreon_acl` (`host_id` , `service_id`,`group_id` ) VALUES ";
         $cpt = 0;
         foreach ($tabGroups as $acl_group_id => $acl_res_id) {
             $tabElem = array();
@@ -441,8 +436,12 @@ try {
             /*
              * Delete old data for this group
              */
-            $DBRESULT = $pearDBO->query("DELETE FROM `centreon_acl` WHERE `group_id` = '" . $acl_group_id . "'");
-            
+            if ($partition == 0) {
+                $DBRESULT = $pearDBO->query("DELETE FROM `centreon_acl` WHERE `group_id` = '" . $acl_group_id . "'");
+            } else {
+
+            }
+
             /** ***********************************************
              * Select
              */
@@ -457,23 +456,15 @@ try {
                 $Host = array();
                 /* ------------------------------------------------------------------ */
                 
-                /*
-                 * Get all Hosts
+                /* ********************************************************************
+                 * Add Hosts 
                  */
-                if (isset($hostIncCache[$res2["acl_res_id"]])) {
-                    foreach ($hostIncCache[$res2["acl_res_id"]] as $host_id => $host_name) {
-                        $Host[$host_id] = $host_name;
-                    }
-                }
-                
                 if (isset($hgResCache[$res2['acl_res_id']])) {
                     foreach ($hgResCache[$res2['acl_res_id']] as $hgId) {
                         if (isset($hostHGRelation[$hgId])) {
                             foreach ($hostHGRelation[$hgId] as $host_id) {
                                 if ($hostCache[$host_id]) {
-                                    $Host[$host_id] = $hostCache[$host_id];
-                                } else {
-                                    print "Host $host_id unknown !\n";
+                                    $Host[$host_id] = 1;
                                 }
                             }
                         }
@@ -481,8 +472,17 @@ try {
                 }
                 
                 if (isset($hostExclCache[$res2["acl_res_id"]])) {
-                    foreach ($hostExclCache[$res2["acl_res_id"]] as $host_id => $host_name) {
+                    foreach ($hostExclCache[$res2["acl_res_id"]] as $host_id => $flag) {
                         unset($Host[$host_id]);
+                    }
+                }
+
+                /*
+                 * Get all Hosts
+                 */
+                if (isset($hostIncCache[$res2["acl_res_id"]])) {
+                    foreach ($hostIncCache[$res2["acl_res_id"]] as $host_id => $flag) {
+                        $Host[$host_id] = 1;
                     }
                 }
                 
@@ -494,15 +494,16 @@ try {
                 /*
                  * get all Service groups
                  */
-                $sgReq = "SELECT host_name, host_id, service_description, service_id 
-                            FROM `acl_resources_sg_relations`, `servicegroup_relation`, `host`, `service`
+                $sgReq = "SELECT host_id, service_id 
+                            FROM acl_resources_sg_relations, servicegroup_relation, host, service
                             WHERE acl_res_id = '" . $res2["acl_res_id"] . "'
                 		        AND host.host_id = servicegroup_relation.host_host_id
                 		        AND service.service_id = servicegroup_relation.service_service_id
                 		        AND servicegroup_relation.servicegroup_sg_id = acl_resources_sg_relations.sg_id
                                 AND service_activate = '1'
                 		  UNION
-                          SELECT host_name, host_id, service_description, service_id FROM `acl_resources_sg_relations`, `servicegroup_relation`, `host`, `service`, `hostgroup`, `hostgroup_relation`
+                          SELECT host_id, service_id 
+                            FROM acl_resources_sg_relations, servicegroup_relation, host, service, hostgroup, hostgroup_relation
                 		    WHERE acl_res_id = '" . $res2["acl_res_id"] . "'
                 		        AND hostgroup.hg_id = servicegroup_relation.hostgroup_hg_id
                                 AND servicegroup_relation.hostgroup_hg_id = hostgroup_relation.hostgroup_hg_id
@@ -515,23 +516,23 @@ try {
                 $tmpH = array();
                 if ($DBRESULT3->numRows()) {
                     while ($h = $DBRESULT3->fetchRow()) {
-                        if (!isset($sgElem[$h["host_name"]])) {
-                            $sgElem[$h["host_name"]] = array();
-                            $tmpH[$h['host_id']] = $h['host_name'];
+                        if (!isset($sgElem[$h["host_id"]])) {
+                            $sgElem[$h["host_id"]] = array();
+                            $tmpH[$h['host_id']] = 1;
                         }
-                        $sgElem[$h["host_name"]][$h["service_description"]] = $h["host_id"] . "," . $h["service_id"];
+                        $sgElem[$h["host_id"]][$h["service_id"]] = $h["host_id"] . "," . $h["service_id"];
                     }
                 }
                 $DBRESULT3->free();
                 
-                foreach ($tmpH as $key => $value) {
-                    $tab = getAuthorizedServicesHost($key, $acl_group_id, $res2["acl_res_id"], $authorizedCategories);
-                    foreach ($tab as $desc => $id) {
-                        if (isset($sgElem[$value]) && isset($sgElem[$value][$desc])) {
-                            if (!isset($tabElem[$value])) {
-                                $tabElem[$value] = array();
+                foreach ($tmpH as $host_id => $flag) {
+                    $tab = getAuthorizedServicesHost($host_id, $acl_group_id, $res2["acl_res_id"], $authorizedCategories);
+                    foreach ($tab as $service_id => $flag2) {
+                        if (isset($sgElem[$value]) && isset($sgElem[$host_id][$service_id])) {
+                            if (!isset($tabElem[$host_id])) {
+                                $tabElem[$host_id] = array();
                             }
-                            $tabElem[$value][$desc] = $key . "," . $id;
+                            $tabElem[$host_id][$service_id] = 1;
                         }
                     }
                     unset($tab);
@@ -546,13 +547,13 @@ try {
                 /*
                  * Initialize and first filter
                  */
-                foreach ($Host as $key => $value) {
-                    $tab = getAuthorizedServicesHost($key, $acl_group_id, $res2["acl_res_id"], $authorizedCategories);
-                    if (!isset($tabElem[$value])) {
-                        $tabElem[$value] = array();
+                foreach ($Host as $host_id => $host_name) {
+                    $tab = getAuthorizedServicesHost($host_id, $acl_group_id, $res2["acl_res_id"], $authorizedCategories);
+                    if (!isset($tabElem[$host_id])) {
+                        $tabElem[$host_id] = array();
                     }
-                    foreach ($tab as $desc => $id) {
-                        $tabElem[$value][$desc] = $key . "," . $id;
+                    foreach ($tab as $desc => $svc_id) {
+                        $tabElem[$host_id][$svc_id] = 1;
                     }
                     unset($tab);
                 }
@@ -569,35 +570,14 @@ try {
                 $str = "";
                 if (count($tabElem)) {
                     $i = 0;
-                    foreach ($tabElem as $host => $svc_list) {
-                        $singleId = array_search($host, $hostCache);
-                        if ($singleId) {
-                            if ($str != "") {
-                                $str .= ", ";
-                            }
-                            $str .= " ({$singleId}, NULL, {$acl_group_id}) ";
+                    foreach ($tabElem as $host_id => $svc_list) {
+                        /* inject host */
+                        $injectHandler->execute(array($host_id, null, $acl_group_id));
+
+                        /* inject service */
+                        foreach ($svc_list as $svc_id => $flag) {
+                            $injectHandler->execute(array($host_id, $svc_id, $acl_group_id));
                         }
-                        foreach ($svc_list as $desc => $t) {
-                            if ($str != "") {
-                                $str .= ', ';
-                            }
-                            $id_tmp = preg_split("/\,/", $t);
-                            $str .= "('" . $id_tmp[0] . "' , '" . $id_tmp[1] . "' , " . $acl_group_id . ") ";
-                            $i++;
-                            if ($i >= 1000) {
-                                $pearDBO->query($strBegin . $str);
-                                $str = "";
-                                $i = 0;
-                            }
-                        }
-                    }
-                    
-                    /*
-                     * Insert datas
-                     */
-                    if ($str != "") {
-                        $pearDBO->query($strBegin . $str);
-                        $str = "";
                     }
                 }
